feat: Entity 연관 관계 설정

+++ src/main/java/com/nhnacademy/springbootjpa/entity/Categories.java
@@ -13,6 +13,6 @@ import lombok.NoArgsConstructor;
 public class Categories {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int categoryId;
+    private long categoryId;
     private String categoryName;
 }
